refactor: move high refresh rate code to prpr engine layer
Move the Android high refresh rate implementation from the phira
application layer (android_hrz.rs) to the prpr engine layer (ext.rs),
following maintainer feedback to put it in prpr-miniquad.

- Add request_high_refresh_rate() to prpr::ext
- Remove phira::android_hrz module
- Update phira to call prpr::ext::request_high_refresh_rate()
