+++ phira/src/lib.rs
@@ -22,7 +22,6 @@ mod tabs;
 mod tags;
 mod threed;
 mod uml;
-mod android_hrz;
 
 use anyhow::Result;
 use data::Data;
@@ -171,7 +170,7 @@ mod dir {
 async fn the_main() -> Result<()> {
     log::register();
 
-    android_hrz::request_high_refresh_rate();
+    prpr::ext::request_high_refresh_rate();
 
     #[cfg(target_env = "ohos")]
     {
